requirements.txt
asgiref==3.3.4
attrs==21.2.0
certifi==2020.12.5
chardet==4.0.0
Django==3.2.3
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.0
flake8==3.9.2
flake8-quotes==3.2.0
idna==2.10
iniconfig==1.1.1
mccabe==0.6.1
packaging==20.9
pluggy==0.13.1
py==1.10.0
pycodestyle==2.7.0
pyflakes==2.3.1
PyJWT==2.1.0
pyparsing==2.4.7
pytest==6.2.4
pytest-django==4.3.0
python-dotenv==0.17.1
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.4
